requirements.txt
nltk
psycopg2-binary
requests
beautifulsoup4
googlesearch-python
